requirements.txt
asgiref==3.4.1
azt==0.0.7
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
Cython==0.29.25
Django==4.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
pycparser==2.21
pyfiglet==0.8.post1
pyzt==1.0.1
sqlparse==0.4.2
termcolor==1.1.0
tzdata==2021.5
